pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zerobp"
version = "0.1.0"
description = "ZeroBP reproduction scaffold (Hydra + PyTorch Lightning)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "lightning>=2.2.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "numpy>=1.24",
    "pytest>=7.4",
    "pillow>=9.5",
    "wandb>=0.16",
]

[tool.setuptools.packages.find]
where = ["src"]